[feat] added a game manager for ground floor
