--- Assets/RoomTransitionScript.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a8cb5e72c4d9d5542bf65c0d442b670f
